Fix for TerritoryWar Superclass
- It seems that, previously, for the participants of the Territory War,
the handler method that is responsible for counting the deaths of the
Quests, was not working well (it did ... fatally).

At the end of the count of deaths in Quest (being in a Territory War),
was re-counting from the beginning, granting in the end more Territory
War points that should be correctly given to the player, as well as the
reward of EXP and SP. For all this at the end of everything, the badges
that are delivered are many more than those that "legally" must be in
RETAIL mode.

The anomaly also produced faults in the remaining Quests associated with
the Territory War in a similar way if it requires the use of the same
handler method.